% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_shares.R
\name{market_shares}
\alias{market_shares}
\title{Compute market shares in an airport or city}
\usage{
market_shares(dataframe, carrier, orig)
}
\arguments{
\item{dataframe}{Data to base the market shares on; should have a Passengers
column}

\item{carrier}{Column name containing the carrier for each ticket, unquoted}

\item{orig}{Column name containing the origin city or airport, unquoted}
}
\value{
a dataframe with the carrier and orig columns, and a market_share
column that contains the market share of that carrier in that location,
as a proportion.
}
\description{
Compute market shares by airport or by city, with flexibility to specify how
the carrier is defined.
}
\details{
This will calculate market shares and return them as proportions by airport
or city and by operator.
}
\examples{
\dontshow{if (require("tibble")) withAutoprint(\{ # examplesIf}
data = tibble::tribble(
  ~origin, ~airline, ~Passengers,
  "CLT",    "American", 10,
  "CLT",    "United",  5,
  "RDU",    "American", 5,
  "RDU",    "United", 7
)

market_shares(data, airline, carrier)
\dontshow{\}) # examplesIf}
}
